requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.9.3
argcomplete==2.0.0
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.1
awscli==2.13.4
awscrt==0.18.0
azure-common==1.1.28
azure-core==1.28.0
azure-identity==1.10.0
azure-mgmt-core==1.3.2
azure-mgmt-rdbms==10.2.0b10
azure-mgmt-resource==23.1.0b2
azure-mgmt-subscription==3.0.0
Babel==2.11.0
bcrypt==3.2.2
Beaker==1.12.1
beautifulsoup4==4.12.2
bidict==0.22.0
blinker==1.5
blivet==3.7.1
blivet-gui==2.4.2
boto3==1.26.111
botocore==1.29.111
Brlapi==0.8.4
Brotli==1.0.9
CacheControl==0.12.11
cached-property==1.5.2
certifi==2022.9.24
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
colorama==0.4.6
crashtest==0.4.1
cryptography==37.0.2
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.6
distro==1.8.0
dnspython==2.3.0
docker==5.0.3
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.19
dulwich==0.21.2
email-validator==1.1.3
evdev==1.6.1
eventlet==0.33.3
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.8.2
Flask==2.2.3
Flask-Babel==2.0.0
flask-compress==1.13
Flask-Cors==4.0.0
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fros==1.1
frozenlist==1.4.0
gpg==1.17.1
greenlet==1.1.3
gssapi==1.7.3
html5lib==1.1
httpagentparser==1.9.5
humanize==3.13.1
idna==3.4
importlib-metadata==6.0.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.3.1
jsonschema==4.17.3
keyring==23.11.0
koji==1.33.0
langtable==0.0.63
ldap3==2.9.1
libcomps==0.1.18
lockfile==0.12.2
lxml==4.9.2
Mako==1.2.3
MarkupSafe==2.1.2
more-itertools==8.13.0
msal==1.22.0
msal-extensions==1.0.0
msgpack==1.0.4
msrest==0.7.1
multidict==6.0.4
nftables==0.1
nltk==3.8.1
numpy==1.24.4
nvautoinstall==0.4.1
oauthlib==3.2.1
olefile==0.46
openai==0.27.8
packaging==23.0
pandas==2.0.3
paramiko==3.1.0
passlib==1.7.4
Paste==3.5.2
pexpect==4.8.0
pid==2.2.3
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==2.6.0
ply==3.11
poetry==1.3.2
poetry-core==1.4.0
poetry-plugin-export==1.3.0
portalocker==2.7.0
productmd==1.36
progressbar2==3.53.2
prompt-toolkit==3.0.36
psutil==5.9.5
psycopg2==2.9.6
ptyprocess==0.7.0
pwquality==1.4.5
pyasn1==0.4.8
pycairo==1.23.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.45.2
pyenchant==3.2.2
pygit2==1.12.2
Pygments==2.14.0
PyGObject==3.44.1
PyJWT==2.6.0
pykickstart==3.47
PyNaCl==1.5.0
pynput==1.7.6
pyOpenSSL==21.0.0
pyotp==2.9.0
pyparsing==3.0.9
pyparted==3.12.0
pyrsistent==0.19.3
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.7.0
python-meh==0.51
python-socketio==5.9.0
python-utils==3.3.3
python-xlib==0.33
pytz==2023.3.post1
pyudev==0.24.0
pyxdg==0.27
PyYAML==6.0
qrcode==7.3.1
ranger-fm==1.9.3
rapidfuzz==2.13.7
redis==4.3.4
regex==2023.6.3
requests==2.28.2
requests-file==1.5.1
requests-ftp==0.3.1
requests-gssapi==1.2.3
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rpm==4.18.1
rpmautospec==0.3.5
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
s3transfer==0.6.2
SecretStorage==3.3.3
selinux==3.5
sepolicy==3.5
setools==4.4.3
shellingham==1.5.0
simpleaudio==1.0.4
simplejson==3.18.3
simpleline==1.9.0
six==1.16.0
sos==4.5.1
soupsieve==2.4.1
SQLAlchemy==1.4.49
sqlparse==0.4.2
sshtunnel==0.4.0
systemd-python==235
Tempita==0.5.2
texttable==1.6.4
tomlkit==0.11.4
tqdm==4.65.0
trash-cli==0.22.10.20
trove-classifiers==2023.2.20
typing_extensions==4.5.0
tzdata==2023.3
ua-parser==0.18.0
urllib3==1.26.16
user-agents==2.2.0
virtualenv==20.21.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.2.3
wrapt==1.14.1
WTForms==3.0.1
yarl==1.9.2
zipp==3.12.0
